==> Notebooks/01_Data_Processing_Team8_Prj4.ipynb <==
import nltk
nltk.download('stopwords')


import pandas as pd
import numpy as np
import re
import nltk
from nltk.corpus import stopwords
from nltk.stem import PorterStemmer,WordNetLemmatizer
from sklearn.model_selection import train_test_split
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.metrics import accuracy_score, confusion_matrix, classification_report

# **Data Pre-Processing**

#mount the drive.
from google.colab import drive
drive.mount('/content/drive')

%ls


%cd /content/drive/My Drive/Group8-Project4/Resources/

# defined path to csv
df = pd.read_csv('WELFake_Dataset.csv')

# display the first few rows of the DataFrame
df.head(10)

#see df info
df.info()

#identify size of table/ data set
df.shape

#drop the Unnamed: 0 column as we will not need it for the analysis
df.drop(['Unnamed: 0'], axis=1, inplace= True)

# Converting the 'title' and 'text' columns to lowercase
df['title'] = df['title'].str.lower()
df['text'] = df['text'].str.lower()

#check for Null values
Null_values = df.isnull().sum()
Null_values

#drop Nan_values
df= df.fillna('')

#check that Nan values have been replaced correctly
df.isnull().sum()

#print the new shape of the dataset
df.shape

# see the distribution of the labels real vs fake
print(f'Ratio of real and fake news:')
df['label'].value_counts(normalize=True).rename({1: 'real', 0: 'fake'})

# combine the title and the text columns
df["title_text"] = df["title"] + df["text"]
df.head(10)

#removing all special characters AND numbers from the title_text column
import re

# create a function to remove punctuation and special characters in case I need the same function later
def remove_special_characters(text):
    return re.sub(r'[^\w\s]|[\d]', '', text)

# Apply the function to the 'title_text' column
df['title_text'] = df['title_text'].apply(remove_special_characters)

#see transformed df
df.head()

#remove stop words using the nltk
stop_words = set(stopwords.words('english'))

# Function to remove stopwords
def remove_stopwords(text):
    words = text.split()
    filtered_words = [word for word in words if word not in stop_words]
    return ' '.join(filtered_words)

# Apply the function to your column
df['title_text'] = df['title_text'].apply(remove_stopwords)


#check update
df.head()

# Creating a new DataFrame with only 'text_title' and 'labels' columns
model_df = df[['title_text', 'label']]
model_df.head()

#Exporting the model_df as a csv for easy retrival if necessary
model_df.to_csv('/content/drive/My Drive/Group8-Project4/Resources/FakeNews_Processed_Data.csv', index=False)
