==> Notebooks/02_fake_news_natural_language_processing.ipynb <==
import pandas as pd
import numpy as np
import re
import nltk
from nltk.corpus import stopwords
from nltk.stem import PorterStemmer,WordNetLemmatizer
from sklearn.model_selection import train_test_split
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.metrics import accuracy_score, confusion_matrix, classification_report

#mount the drive.
from google.colab import drive
drive.mount('/content/drive')

#naviagte to directory where clean csv is
%cd /content/drive/My Drive/Group8-Project4/Resources/

%ls

# defined path to csv
df = pd.read_csv('FakeNews_Processed_Data.csv')

# display the first few rows of the DataFrame
df.head(10)

# Print the data type of each entry in 'title_text'
print(df['title_text'].apply(type).value_counts())


# Convert entire 'title_text' column to strings
df['title_text'] = df['title_text'].astype(str)



## **Word Embedding**
We will use Word2Vec for word embedding. We will use a pre-trained model to vectorise our data to start with because these models are trained on a massive corpus and can capture a wide range of semantic relationships. We will access the Google Word2Vec model.

!pip install gensim


# Word2Vec expects a list of tokenized sentences, where each sentence is a list of words
#this code is tokenizing the text

from nltk.tokenize import word_tokenize
nltk.download('punkt')  # Download the tokenizer model

tokenized_data = [word_tokenize(sentence) for sentence in df['title_text']]

%cd /content/drive/My Drive/Group8-Project4


%ls

# Adjust the path according to your Google Drive structure
tokenized_data_path = '/content/drive/My Drive/Group8-Project4/tokenized_data.pkl'


import pickle

# Saving the tokenized_data
with open(tokenized_data_path, 'wb') as file:
    pickle.dump(tokenized_data, file)


#accessing the saved tokenized_data
tokenized_data_path = '/content/drive/My Drive/Group8-Project4/tokenized_data.pkl'


import pickle

# Load the tokenized data
with open(tokenized_data_path, 'rb') as file:
    loaded_tokenized_data = pickle.load(file)


#check that the loading was succesfull

print(loaded_tokenized_data[:2])  # print first two tokenized documents
print(len(loaded_tokenized_data)) # print the number of documents


#using the google pre-trained word2Vec model
import gensim.downloader as api

# Load Google's pre-trained Word2Vec model.
w2v_model = api.load('word2vec-google-news-300')


import numpy as np

def document_vector(word_list, model):
    # remove out-of-vocabulary words and get their vectors
    word_vectors = [model[word] for word in word_list if word in model.key_to_index]

    if len(word_vectors) == 0:
        return np.zeros(model.vector_size)  # return zero vector if no words in the model
    else:
        # calculate the mean of word vectors to represent the document vector
        return np.mean(word_vectors, axis=0)


# Vectorize each document in the tokenized data
X = np.array([document_vector(doc, w2v_model) for doc in tokenized_data])


print(X.shape)

# Print the first vector to inspect
print(X[0])


%ls

%cd  /content/drive/My Drive/Group8-Project4

%ls

# Define the file path for the numpy file
npy_path = '/content/drive/My Drive/Group8-Project4/vectorized_data.npy'


# Save the array to a .npy file
np.save(npy_path, X)
