==> Notebooks/03_Logistic_Regression_Model#1.ipynb <==
import pandas as pd
import numpy as np


from google.colab import drive
drive.mount('/content/drive')


%ls

# Define paths to point to the correct location in Google Drive
csv_path = '/content/drive/My Drive/Group8-Project4/Resources/FakeNews_Processed_Data.csv'
npy_path = '/content/drive/My Drive/Group8-Project4/vectorized_data.npy'


# Load the data frame
df = pd.read_csv(csv_path)

df.head(10)

# Extract labels
y = df['label'].values


# Load vectorized features
X = np.load(npy_path)

#check y
print(y)

#checking that X has loaded properly
print(X[0])

# **Split our Data into Training and Testing**

from sklearn.model_selection import train_test_split

# Split the data - 80% for training, 20% for testing
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42, stratify=y)


# **Create a logistic Regression Model**

from sklearn.linear_model import LogisticRegression
classifier = LogisticRegression(max_iter=1000, random_state=1)
classifier

# **Fit training data to our model**

# Train the data
classifier.fit(X_train, y_train)

# **Validate the model using test data**

# Score the model
print(f"Training Data Score: {classifier.score(X_train, y_train)}")
print(f"Testing Data Score: {classifier.score(X_test, y_test)}")

# **Make Predictions**

 # Predict outcomes for test data set
predictions = classifier.predict(X_test)
pd.DataFrame({"Prediction": predictions, "Actual": y_test})

from sklearn.metrics import accuracy_score
# Display the accuracy score for the test dataset.
accuracy_score(y_test, predictions)

# **Optimsing the Model**

#define parameter grid
param_grid = {
    'C': np.logspace(-4, 4, 20),  # A range of regularization strengths
    'penalty': ['l1', 'l2'],  # L1 and L2 regularization
    'solver': ['liblinear', 'saga']  # Solvers that can handle both L1 and L2 penalty
}


#Initialize Logistic Regression and GridSearchCV
from sklearn.model_selection import GridSearchCV
from sklearn.linear_model import LogisticRegression
log_reg = LogisticRegression()
grid_search = GridSearchCV(log_reg, param_grid, cv=5, verbose=0, n_jobs=-1)

#Fit the Grid Search to the Data
grid_search.fit(X_train, y_train)


print("Best parameters:", grid_search.best_params_)
print("Best score:", grid_search.best_score_)


#Use the best model found by the grid search to make predictions or further analyze
best_model = grid_search.best_estimator_
y_pred = best_model.predict(X_test)


#Evaluate the model on test data
from sklearn.metrics import accuracy_score, classification_report

print("Test set accuracy:", accuracy_score(y_test, y_pred))
print(classification_report(y_test, y_pred))


# **Save the Model Using joblib**

!pip install joblib


from joblib import dump

# Define the path in your Google Drive to save the model
model_path = '/content/drive/My Drive/Group8-Project4/my_logistic_regression_model.joblib'

# Save your model
dump(best_model, model_path)


#Loading the Model later
# from joblib import load

# # Same path where you saved it
# model_path = '/content/drive/My Drive/Group8-Project4/my_logistic_regression_model.joblib'

# # Load the model
# loaded_model = load(model_path)
