==> Notebooks/04_RandomForrest_Model.ipynb <==
import numpy as np
import pandas as pd
from sklearn.ensemble import RandomForestClassifier
from sklearn.model_selection import train_test_split
from sklearn.metrics import accuracy_score, classification_report


#mount google drive
from google.colab import drive
drive.mount('/content/drive')

# load label from csv
df = pd.read_csv('/content/drive/My Drive/Group8-Project4/Resources/FakeNews_Processed_Data.csv')
y = df['label'].values


#load vectorised features

X = np.load('/content/drive/My Drive/Group8-Project4/vectorized_data.npy')


#split data into training and testing sets

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)


#initialise and train the random forest classifier

rf_classifier = RandomForestClassifier(n_estimators=500, random_state=42)

# Train the classifier
rf_classifier.fit(X_train, y_train)


#make predictions adn evaluate the model
#Predict on the testing set
y_pred = rf_classifier.predict(X_test)


# Print the accuracy and classification report
print("Accuracy on test data:", accuracy_score(y_test, y_pred))
print(classification_report(y_test, y_pred))


#save my model to my google drive
model_path = '/content/drive/My Drive/Group8-Project4/my_random_forest_model.joblib'


from joblib import dump

# Assume rfc_best is your trained model
dump(rf_classifier, model_path)


#to load the model use the code below:
# from joblib import load

# # Load the model
# loaded_model = load(model_path)
