==> Notebooks/05_Deep_Machine_Learning_Model.ipynb <==
import pandas as pd
import numpy as np
import tensorflow as tf
from tensorflow.keras.models import Sequential
from tensorflow.keras.layers import Dense, Dropout


from google.colab import drive
drive.mount('/content/drive')

# load label from csv
df = pd.read_csv('/content/drive/My Drive/Group8-Project4/Resources/FakeNews_Processed_Data.csv')
y = df['label'].values


#load vectorised features

X = np.load('/content/drive/My Drive/Group8-Project4/vectorized_data.npy')

from sklearn.model_selection import train_test_split


# Split the data
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)


#build the model
# Initialize the constructor
model = Sequential()

# Add an input layer
model.add(Dense(128, activation='relu', input_shape=(X_train.shape[1],)))

# Add one hidden layer
model.add(Dense(64, activation='relu'))

# Add an output layer
model.add(Dense(1, activation='sigmoid'))


#compile the model
model.compile(loss='binary_crossentropy',
              optimizer='adam',
              metrics=['accuracy'])


#train the model
history = model.fit(X_train, y_train, epochs=20, batch_size=32, validation_split=0.2, verbose=1)


#evaluate the Model
loss, accuracy = model.evaluate(X_test, y_test, verbose=0)
print("Testing Accuracy: {:.4f}".format(accuracy))


#make predictions
predictions = model.predict(X_test)


# **Second Deep Machine Learning Model**

#build the model
# Initialize the constructor
model_2 = Sequential()

# Add an input layer
model_2.add(Dense(128, activation='relu', input_shape=(X_train.shape[1],)))

# Add one hidden layer
model_2.add(Dense(64, activation='sigmoid'))

# Add an output layer
model_2.add(Dense(1, activation='sigmoid'))

#compile the model
model_2.compile(loss='binary_crossentropy',
              optimizer='adam',
              metrics=['accuracy'])


#train the model
history_2 = model_2.fit(X_train, y_train, epochs=20, batch_size=32, validation_split=0.2, verbose=1)

#evaluate the Model
loss, accuracy = model_2.evaluate(X_test, y_test, verbose=0)
print("Testing Accuracy: {:.4f}".format(accuracy))

predictions = model_2.predict(X_test)


print(predictions)

#save the best model
model_path = '/content/drive/My Drive/Group8-Project4/my_deep_learning_model.keras'


model_2.save(model_path)


# #CODE TO USE TO LOAD THE MODEL
# from tensorflow.keras.models import load_model

# # Load the model
# loaded_model = load_model(model_path)
